requirements.txt
-i https://pypi.org/simple
asgiref==3.9.2; python_version >= '3.9'
django==4.2.25; python_version >= '3.8'
django-annoying==0.10.8; python_version >= '3.0'
django-crontab==0.7.1
django-mathfilters==1.0.0
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
faker==37.6.0; python_version >= '3.9'
pillow==11.3.0; python_version >= '3.9'
pyjwt==2.10.1; python_version >= '3.9'
pymysql==1.1.2; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.2
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.15.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
